+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WinForms/FormsPlotExtensions.cs
@@ -2,7 +2,6 @@
 using System.Drawing;
 using System.IO;
 using System.Windows.Forms;
-using ScottPlot.Control;
 
 namespace ScottPlot.WinForms;
 
@@ -50,11 +49,6 @@ public static class FormsPlotExtensions
         processor.Process(action);
     }
 
-    [Obsolete("Double-clicks do not require processing. They are inferred from delay between single clicks.", true)]
-    public static void ProcessDoubleClick(this Interactivity.UserInputProcessor processor, EventArgs e)
-    {
-    }
-
     public static void ProcessMouseWheel(this Interactivity.UserInputProcessor processor, MouseEventArgs e)
     {
         Pixel mousePixel = new(e.X, e.Y);
@@ -67,6 +61,13 @@ public static class FormsPlotExtensions
     }
 
     public static void ProcessKeyDown(this Interactivity.UserInputProcessor processor, KeyEventArgs e)
+    {
+        Interactivity.Key key = e.GetKey();
+        Interactivity.IUserAction action = new Interactivity.UserActions.KeyDown(key);
+        processor.Process(action);
+    }
+
+    public static void ProcessKeyDown(this Interactivity.UserInputProcessor processor, PreviewKeyDownEventArgs e)
     {
         Interactivity.Key key = e.GetKey();
         Interactivity.IUserAction action = new Interactivity.UserActions.KeyDown(key);
@@ -86,6 +87,11 @@ public static class FormsPlotExtensions
     }
 
     internal static Interactivity.Key GetKey(this KeyEventArgs e)
+    {
+        return GetKey(e.KeyCode);
+    }
+
+    internal static Interactivity.Key GetKey(this PreviewKeyDownEventArgs e)
     {
         return GetKey(e.KeyCode);
     }
@@ -121,28 +127,6 @@ public static class FormsPlotExtensions
             : new Interactivity.Key($"Unknown modifier key {keyName}");
     }
 
-    internal static Control.MouseButton Button(this MouseEventArgs e)
-    {
-        return e.Button switch
-        {
-            System.Windows.Forms.MouseButtons.Left => Control.MouseButton.Left,
-            System.Windows.Forms.MouseButtons.Right => Control.MouseButton.Right,
-            System.Windows.Forms.MouseButtons.Middle => Control.MouseButton.Middle,
-            _ => Control.MouseButton.Unknown,
-        };
-    }
-
-    internal static Key Key(this KeyEventArgs e)
-    {
-        return e.KeyCode switch
-        {
-            Keys.ControlKey => Control.Key.Ctrl,
-            Keys.Menu => Control.Key.Alt,
-            Keys.ShiftKey => Control.Key.Shift,
-            _ => Control.Key.Unknown,
-        };
-    }
-
     internal static Bitmap GetBitmap(this Plot plot, int width, int height)
     {
         byte[] bytes = plot.GetImage(width, height).GetImageBytes();
@@ -155,11 +139,5 @@ public static class FormsPlotExtensions
     {
         using MemoryStream ms = new(img.GetImageBytes(ImageFormat.Bmp));
         return new Bitmap(ms);
-    }
-
-    public static void CopyToClipboard(this SavedImageInfo info)
-    {
-        System.Drawing.Bitmap bmp = new(info.Path);
-        Clipboard.SetImage(bmp);
     }
 }