+++ src/ScottPlot5/ScottPlot5/Interactivity/UserActionResponses/MouseDragPan.cs
@@ -31,21 +31,30 @@ public class MouseDragPan(MouseButton button) : IUserActionResponse
     /// </summary>
     public bool LockX { get; set; } = false;
 
-    public void ResetState(Plot plot)
+    /// <summary>
+    /// If enabled, mouse interactions over a single axis will be applied to all axes with the same orientation.
+    /// </summary>
+    public bool ChangeOpposingAxesTogether { get; set; } = false;
+
+    public void ResetState(IPlotControl plotControl)
     {
         RememberedLimits = null;
         MouseDownPixel = Pixel.NaN;
     }
 
-    public ResponseInfo Execute(Plot plot, IUserAction userInput, KeyboardState keys)
+    public ResponseInfo Execute(IPlotControl plotControl, IUserAction userInput, KeyboardState keys)
     {
         // mouse down starts drag
         if (userInput is IMouseButtonAction mouseDownAction
             && mouseDownAction.Button == MouseButton
             && mouseDownAction.IsPressed)
         {
-            MouseDownPixel = mouseDownAction.Pixel;
-            RememberedLimits = new(plot);
+            Plot? plot = plotControl.GetPlotAtPixel(mouseDownAction.Pixel);
+            if (plot is not null)
+            {
+                RememberedLimits = new(plot);
+                MouseDownPixel = mouseDownAction.Pixel;
+            }
 
             return ResponseInfo.NoActionRequired;
         }
@@ -57,6 +66,7 @@ public class MouseDragPan(MouseButton button) : IUserActionResponse
             && RememberedLimits is not null)
         {
             RememberedLimits.Recall();
+            Plot plot = RememberedLimits.Plot;
             RememberedLimits = null;
             ApplyToPlot(plot, MouseDownPixel, mouseUpAction.Pixel, keys);
 
@@ -78,6 +88,7 @@ public class MouseDragPan(MouseButton button) : IUserActionResponse
             }
 
             RememberedLimits.Recall();
+            Plot plot = RememberedLimits.Plot;
             ApplyToPlot(plot, MouseDownPixel, mouseAction.Pixel, keys);
             return new ResponseInfo() { RefreshNeeded = true, IsPrimary = true };
         }
@@ -97,6 +108,6 @@ public class MouseDragPan(MouseButton button) : IUserActionResponse
             px2.Y = px1.Y;
         }
 
-        MouseAxisManipulation.DragPan(plot, px1, px2);
+        MouseAxisManipulation.DragPan(plot, px1, px2, ChangeOpposingAxesTogether);
     }
 }