+++ src/ScottPlot5/ScottPlot5/Interfaces/IMultiplotLayout.cs
@@ -2,5 +2,5 @@
 
 public interface IMultiplotLayout
 {
-    IEnumerable<(FractionRect, Plot)> GetLayout(IReadOnlyList<Plot> plots);
+    void ResetAllPositions(Multiplot multiplot);
 }