+++ src/ScottPlot5/ScottPlot5/Plot.cs
@@ -1,9 +1,7 @@
 ﻿using ScottPlot.AxisPanels;
-using ScottPlot.Control;
 using ScottPlot.Grids;
 using ScottPlot.Rendering;
 using ScottPlot.Stylers;
-using System.ComponentModel;
 
 namespace ScottPlot;
 
@@ -35,7 +33,7 @@ public class Plot : IDisposable
     public PlotStyler Style { get; } = new();
 
     public FontStyler Font { get; }
-    public Legend Legend { get; }
+    public Legend Legend { get; set; }
 
     public DefaultGrid Grid => Axes.DefaultGrid;
 
@@ -49,9 +47,19 @@ public class Plot : IDisposable
     public object Sync { get; } = new();
 
     /// <summary>
-    /// The user control this plot belongs to
+    /// Wire this plot to be aware of the interactive control it is part of
     /// </summary>
-    public IPlotControl? PlotControl { get; set; } = null;
+    public IPlotControl? PlotControl
+    {
+        set
+        {
+            ParentControl = value;
+        }
+    }
+
+    private IPlotControl? ParentControl = null;
+
+    public bool HasParentControl => ParentControl is not null;
 
     public Plot()
     {
@@ -61,7 +69,7 @@ public class Plot : IDisposable
         RenderManager = new(this);
         Legend = new(this);
         Layout = new(this);
-        ZoomRectangle = new StandardZoomRectangle(this);
+        ZoomRectangle = new Plottables.ZoomRectangle(this);
     }
 
     public void Dispose()
@@ -280,11 +288,11 @@ public class Plot : IDisposable
     /// <summary>
     /// Render the plot and return an HTML img element containing a Base64-encoded PNG
     /// </summary>
-    public string GetPngHtml(int width, int height)
+    public string GetPngHtml(int width, int height, string classContent = "", string styleContent = "")
     {
         Image img = GetImage(width, height);
         byte[] bytes = img.GetImageBytes();
-        return ImageOperations.GetImageHtml(bytes);
+        return ImageOperations.GetImageHtml(bytes, "png", classContent, styleContent);
     }
 
     /// <summary>
@@ -675,6 +683,22 @@ public class Plot : IDisposable
         {
             panel.ShowDebugInformation = enable;
         }
+    }
+
+    #endregion
+
+    #region Interactivity
+
+    // These methods were introduced to help reduce reliance on deprecated IPlotControl.Interaction and Plot.PlotControl classes
+
+    internal void ParentControlRefresh()
+    {
+        ParentControl?.Refresh();
+    }
+
+    internal void ParentControlShowContextMenu(Pixel position)
+    {
+        ParentControl?.ShowContextMenu(position);
     }
 
     #endregion