+++ CHANGELOG.md
@@ -1,5 +1,24 @@
-## ScottPlot 5.0.47
+## ScottPlot 5.0.48
 _Not yet on NuGet..._
+* Multiplot: Extend `Multiplot` to support rending onto an existing surface (#4628, #4600)
+* User Controls: Add `Multiplot` to `IPlotControl` (#4628, #4600)
+* User Controls: Remove deprecated `Interaction` system (#4630, #4600)
+* Plot: Remove the `PlotControl` property to encourage one-way flow of control (#4631, #4600)
+* Controls: Update all `UserInputProcessor` code to support `Multiplot` interactions (#4632, #4600)
+* Multiplot: Add advanced subplot positioning features (#4633, #4600)
+* Multiplot: Add shared axis helper methods to facilitate syncing axis limits across subplots (#4635, #4600)
+* Generate:  Fixed causing `AddNoiseInPlace()` to offset the mean of the underlying signal (#4564) @nightroman
+* Text: Fixed `DragTo()` to use the correct axis dimension (#4587) @claire0821
+* Rendering: Improved handling of the `RenderPack` inside the render loop to avoid Skia exceptions (#4592) @vadimart92
+* Axes: Improved support for `CoordinateRect` hit detection on plots with inverted axes (#4596) @SprinterDave
+* Axes: Improved axis rule behavior on plots using inverted axes (#3597, #4609) @CoderPM2011
+* Smith Chart: New plot type for displaying impedance of signals as resistance vs. reactance on a circular axis system (#4634, #4578, #41) @CoderPM2011 @louis1901 @Padanian
+* Rendering: Improve support for platforms where BGRA is not the default color model (#4648, #4647) @bclehmann @ssharks
+* Legend: Made `GetItems()` virtual to support custom ordering logic and made `Plot.Legend` settable (#4571) @onur-akaydin
+* Controls: improved middle-click-drag zoom rectangle support for plots using inverted axis limits (#4573) @xichaoqiang
+
+## ScottPlot 5.0.47
+_Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-11-24_
 * Controls: Fix issue preventing the context menu from appearing after it was used to open a new window (#4529) @david3951445
 * Interactivity: Created `HitablePlottableDecorator` and `DragablePlottableDecorator` classes that wrap any `IPlottable` to add pixel-based mouse collision detection and drag capability to any plot type (#4531, #4496) @StendProg
 * Ticks: Created a plottable for displaying multiplier notation and added the `Plot.Axes.SetupMultiplierNotation()` helper method for rapidly enabling it with typical options (#4530) @Paraplegia
@@ -9,6 +28,10 @@ _Not yet on NuGet..._
 * Bar: Created `Add.Ranges()` and `Add.StackedRanges()` to simplify creation of stacked range charts (#4548) @quantfreedom @wellcaffeinated
 * Bracket: New plot type for annotating ranges of linear lines in coordinate space (#4547, #1863) @FULL69 @bclehmann
 * Financial Charting: Improved `FinancialTimeAxis` tick generation behavior (#4483, #4551, #4385) @VladislavPustovarov @quantfreedom
+* Histogram: Created `Add.Histogram()` which accepts a `Histogram` and returns a `HistogramBars` plottable designed for displaying continuously updated histogram counts (#4557) @jpgarza93
+* Legend: Added support for paring each `LegendItem` with its parent `IPlottable` to facilitate mouse interaction (#4533) @BambOoxX
+* Plot: Modified `GetPngHtml()` to add optional `classContent` and `styleContent` arguments to allow CSS style injection
+* Reporting: New `ScottPlot.Reporting` classes support creating multi-figure reports which may be viewed interactively or exported as HTML documents (#4526)
 
 ## ScottPlot 5.0.46
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-11-17_